Update on "Fix signed integer overflow in rng_test.h"
Differential Revision: [D20978925](https://our.internmc.facebook.com/intern/diff/D20978925)

[ghstack-poisoned]

+++ aten/src/ATen/test/rng_test.h
@@ -84,8 +84,8 @@ void test_random_from_to(const at::Device& device) {
             full_64_bit_range_case_covered = true;
           } else {
             if (to.has_value()) {
-              range = *to - from;
-              // range = static_cast<uint64_t>(*to) - static_cast<uint64_t>(from);
+              // range = *to - from; // runtime error: signed integer overflow: 0 - -9223372036854775808 cannot be represented in type 'long'
+              range = static_cast<uint64_t>(*to) - static_cast<uint64_t>(from);
               from_to_case_covered = true;
             } else {
               range = max_val - from + 1;
